Agregando modelos para Ordenes, nuevo botón "Mis compras" en el header
Implementación de ordenes al confirmar la compra. Ahora el usuario puede ver los productos que compró en la página "myShoppings". Esto solo aplica para los productos de tipo Stock, para los Bids se hará una implementación diferente

+++ store/models.py
@@ -34,8 +34,8 @@ class Product(models.Model):
 
 
 class BidProduct(Product):
-    currentBid = models.PositiveIntegerField(blank=True,default=0)
-    minBid = models.PositiveIntegerField(default=0)
+    currentBid = models.FloatField(blank=True,default=0)
+    minBid = models.FloatField(default=0)
     condition = models.CharField(max_length=50)
     bidWinner = models.OneToOneField(User, related_name="BidWinner", null=True, on_delete=models.SET_NULL, blank=True)
